Implementation of Base Health Interface
Creation of the health interface script as 'IHealth', including the variables and functions. Implementation of the 'Health' script which briefly summed is the health of objects. 'TakeDamage' functionality needs shield recharge components added. The 'Heal' function works as intended, healing the player upon collision with the health pickup. Additional components are needed for the recharge and depletion functions.