PickupSpawner Implemented
Created the script for the pickupSpawner which currently spawns the health pickup, a similar 'Test_HealthPickup' to the large one currently within the scene however this one is a prefab. I also added a simple rotation script which just rotates the object at a specified speed.

To implement the ammo pickup and weapon pickup to the spawner we'd need to create both pickups and their respective variables. However it's pretty straight forward on adding them into the spawner from there