AmmoPickup Implemented
The ammopickup works as intended, adding ammo (AddAmmo) to the ammo amount (ammoAmount). This amount is specified in engine. We ran into some issues within the ammoPickup where activeWeapon and other components weren't being recognised. To fix the issues we ended up storing the references and index in variables. Comments within the AmmoPickup scripts will elaborate.