Weapon Pickup with issue
Weapon Pickup now checks to see if the weapon to pick up is stored on the player or not, tells the player to add the weapon, and registers the weapon in the array.

The issue is that we cannot yet choose which type of weapon the pickup is for.  It is currently using the machine gun as the basis.